# backend/pkg/http/controllers/user_controller.go
package controllers

import (
	"github.com/gofiber/fiber/v2"
	"github.com/yzaimoglu/flathunter/pkg/http/services"
	"github.com/yzaimoglu/flathunter/pkg/models"
)

// GetUsers retrieves all the users
func GetUsers(c *fiber.Ctx) error {
	id := c.Params("id")
	user, err := services.GetUser(id)

	if err != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"user": user,
	})
}

// GetUsers retrieves all the users
func InsertUser(c *fiber.Ctx) error {
	insertedId, err := services.InsertUser(models.CreateUser{
		Email:    "yagi@mitocho.io",
		Password: "123456",
	})
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"insertedId": insertedId,
	})
}